build: Update apigeecli to version 0.51.1 or greater
This commit updates the Dockerfile to install apigeecli version 0.51.1 or greater, which is required for the latest features and bug fixes. This will ensure that our Docker images use the latest stable version of apigeecli and will be compatible with future changes to the apigeecli tool. The Dockerfile was also cleaned up to have more efficient layers, reduce the size of the image and make it quicker to build. No functionality changes were made in this commit.